+++ doc/changelog.rst
@@ -23,6 +23,7 @@ Localization
 Admin
 ~~~~~
 
+- Quick add staff members for shops
 - Main menu is now updateable through provides.
 - Add new provide category called `order_printouts_delivery_extra_fields`
   which can be used to add extra rows to order delivery slip.